pyproject.toml
[project]
name = "robust-fastapi-api"
version = "0.1.0"
description = ""
authors = [
    {name = "João Pedro",email = "sousajoaopedro1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "asyncpg (>=0.31.0,<0.32.0)",
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "redis (>=7.2.1,<8.0.0)",
    "structlog (>=25.5.0,<26.0.0)"
]

[tool.poetry]
packages = [{include = "robust_fastapi_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.4,<0.16.0)",
    "testcontainers (>=4.14.1,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501", "PLR0913", "PLR0917", "PLR6301"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format '
pre_build = 'task format'
build = 'docker compose up --build'
pre_dev = 'task format && docker compose up -d database redis'
dev = 'poetry run uvicorn --host 0.0.0.0 --port 8080 robust_fastapi_api.app:app --reload --reload-include "settings/*.yaml" --reload-include ".env"'
pre_test = 'task format'
test = 'pytest -s -x --cov=robust_fastapi_api -vv --cov-fail-under=90'
post_test = 'coverage html'